+++ test/modifiableInstruction.test.ts
@@ -28,4 +28,26 @@ describe("Modifiable Instruction", () => {
         instruction = dockerfile.getInstructions()[0] as ModifiableInstruction;
         assert.equal(instruction.getFlags().length, 0);
     });
+
+    it("toString", () => {
+        let dockerfile = DockerfileParser.parse("HEALTHCHECK CMD ls");
+        let instruction = dockerfile.getInstructions()[0] as ModifiableInstruction;
+        assert.equal(instruction.toString(), "HEALTHCHECK CMD ls");
+
+        dockerfile = DockerfileParser.parse("HEALTHCHECK CMD ls --help");
+        instruction = dockerfile.getInstructions()[0] as ModifiableInstruction;
+        assert.equal(instruction.toString(), "HEALTHCHECK CMD ls --help");
+
+        dockerfile = DockerfileParser.parse("HEALTHCHECK --interval=30s CMD ls");
+        instruction = dockerfile.getInstructions()[0] as ModifiableInstruction;
+        assert.equal(instruction.toString(), "HEALTHCHECK --interval=30s CMD ls");
+
+        dockerfile = DockerfileParser.parse("HEALTHCHECK --interval=30s --timeout=10s CMD ls");
+        instruction = dockerfile.getInstructions()[0] as ModifiableInstruction;
+        assert.equal(instruction.toString(), "HEALTHCHECK --interval=30s --timeout=10s CMD ls");
+
+        dockerfile = DockerfileParser.parse("COPY . .");
+        instruction = dockerfile.getInstructions()[0] as ModifiableInstruction;
+        assert.equal(instruction.toString(), "COPY . .");
+    });
 });